passed verifier

+++ pkg/network/ebpf/c/protocols/http2-maps-defs.h
@@ -2,7 +2,7 @@
 #define __HTTP2_MAPS_DEFS_H
 
 // http2_static_table is the map that holding the supported static values by index and its static value.
-BPF_HASH_MAP(http2_static_table, u64, static_table_entry_t, 20)
+BPF_HASH_MAP(http2_static_table, u8, static_table_entry_t, 20)
 
 // http2_dynamic_table is the map that holding the supported dynamic values - the index is the static index and the
 // tcp_con and it is value is the buffer which contains the dynamic string.
